requirements.txt
flask
numpy
scikit-learn
opencv-python-headless
joblib
pillow
gunicorn
